fix types

+++ dam/tstamp.cpp
@@ -30,13 +30,13 @@ pthread_mutex_t m_tstamp_lock;
 
 static inline uint32_t delta_nsec(const struct timespec *t1, const struct timespec *t0) {
 	uint32_t dsec = t1->tv_sec - t0->tv_sec;
-	uint32_t dnsec = t1->tv_nsec - t0->tv_nsec;
+	int32_t dnsec = t1->tv_nsec - t0->tv_nsec;
 	if (dsec > 0) {
 		printf("more than one second diff. diff: %d\n", dnsec);
 		dnsec += 1000000000;
 		printf("more than one second diff. diff: %d\n", dnsec);
 	}
-	return dnsec; 
+	return u_int32_t(dnsec); 
 }
 
 static inline void delta_time(const struct timespec *t1, const struct timespec *t0, struct timespec *dt) {
